Move all events in a dedicated `events/` folder.
