+++ editor/views/metadata.py
@@ -3,7 +3,8 @@ import enum
 import streamlit as st
 
 from core.state import Metadata
-from utils import needed_field
+from events.metadata import handle_metadata_change
+from events.metadata import MetadataEvent
 
 # List from https://www.kaggle.com/discussions/general/116302.
 licenses = [
@@ -23,29 +24,6 @@ licenses = [
     "CC BY-NC-SA",
     "CC BY-NC-ND",
 ]
-
-
-class MetadataEvent(enum.Enum):
-    """Event that triggers a metadata change."""
-
-    NAME = "NAME"
-    DESCRIPTION = "DESCRIPTION"
-    URL = "URL"
-    LICENSE = "LICENSE"
-    CITATION = "CITATION"
-
-
-def handle_metadata_change(event: MetadataEvent, metadata: Metadata, key: str):
-    if event == MetadataEvent.NAME:
-        metadata.name = st.session_state[key]
-    elif event == MetadataEvent.DESCRIPTION:
-        metadata.description = st.session_state[key]
-    elif event == MetadataEvent.LICENSE:
-        metadata.license = st.session_state[key]
-    elif event == MetadataEvent.CITATION:
-        metadata.citation = st.session_state[key]
-    elif event == MetadataEvent.URL:
-        metadata.url = st.session_state[key]
 
 
 def render_metadata():
