+++ editor/views/source.py
@@ -4,8 +4,11 @@ from typing import Any
 import streamlit as st
 
 from core.state import Field
-from core.state import Metadata
 from core.state import RecordSet
+from events.fields import ExtractType
+from events.fields import FieldEvent
+from events.fields import handle_field_change
+from events.fields import TransformType
 import mlcroissant as mlc
 from utils import DF_HEIGHT
 from utils import needed_field
@@ -16,19 +19,6 @@ class SourceType:
 
     DISTRIBUTION = "distribution"
     FIELD = "field"
-
-
-class ExtractType:
-    """The type of extraction to perform."""
-
-    COLUMN = "Column"
-    JSON_PATH = "JSON path"
-    FILE_CONTENT = "File content"
-    FILE_NAME = "File name"
-    FILE_PATH = "File path"
-    FILE_FULLPATH = "Full path"
-    FILE_LINES = "Lines in file"
-    FILE_LINE_NUMBERS = "Line numbers in file"
 
 
 EXTRACT_TYPES = [
@@ -43,17 +33,6 @@ EXTRACT_TYPES = [
 ]
 
 
-class TransformType:
-    """The type of transformation to perform."""
-
-    FORMAT = "Apply format"
-    JSON_PATH = "Apply JSON path"
-    REGEX = "Apply regular expression"
-    REPLACE = "Replace"
-    SEPARATOR = "Separator"
-
-
-# TODO(marcenacp): Possible to remove?
 TRANSFORM_TYPES = [
     TransformType.FORMAT,
     TransformType.JSON_PATH,
@@ -126,117 +105,6 @@ def _handle_remove_reference(field):
     field.references = mlc.Source()
 
 
-class ChangeEvent(enum.Enum):
-    """Event that triggers a field change."""
-
-    NAME = "NAME"
-    DESCRIPTION = "DESCRIPTION"
-    DATA_TYPE = "DATA_TYPE"
-    SOURCE = "SOURCE"
-    SOURCE_EXTRACT = "SOURCE_EXTRACT"
-    SOURCE_EXTRACT_COLUMN = "SOURCE_EXTRACT_COLUMN"
-    SOURCE_EXTRACT_JSON_PATH = "SOURCE_EXTRACT_JSON_PATH"
-    TRANSFORM = "TRANSFORM"
-    TRANSFORM_FORMAT = "TRANSFORM_FORMAT"
-    REFERENCE = "REFERENCE"
-    REFERENCE_EXTRACT = "REFERENCE_EXTRACT"
-    REFERENCE_EXTRACT_COLUMN = "REFERENCE_EXTRACT_COLUMN"
-    REFERENCE_EXTRACT_JSON_PATH = "REFERENCE_EXTRACT_JSON_PATH"
-
-
-def _get_source(source: mlc.Source | None, value: Any) -> mlc.Source:
-    if not source:
-        source = mlc.Source(extract=mlc.Extract())
-    if value == ExtractType.COLUMN:
-        source.extract = mlc.Extract(column="")
-    elif value == ExtractType.FILE_CONTENT:
-        source.extract = mlc.Extract(file_property=mlc.FileProperty.content)
-    elif value == ExtractType.FILE_NAME:
-        source.extract = mlc.Extract(file_property=mlc.FileProperty.filename)
-    elif value == ExtractType.FILE_PATH:
-        source.extract = mlc.Extract(file_property=mlc.FileProperty.filepath)
-    elif value == ExtractType.FILE_FULLPATH:
-        source.extract = mlc.Extract(file_property=mlc.FileProperty.fullpath)
-    elif value == ExtractType.FILE_LINES:
-        source.extract = mlc.Extract(file_property=mlc.FileProperty.lines)
-    elif value == ExtractType.FILE_LINE_NUMBERS:
-        source.extract = mlc.Extract(file_property=mlc.FileProperty.lineNumbers)
-    elif value == ExtractType.JSON_PATH:
-        source.extract = mlc.Extract(json_path="")
-    return source
-
-
-def handle_field_change(
-    change: ChangeEvent,
-    field: Field,
-    key: str,
-    **kwargs,
-):
-    value = st.session_state[key]
-    if change == ChangeEvent.NAME:
-        field.name = value
-    elif change == ChangeEvent.DESCRIPTION:
-        field.description = value
-    elif change == ChangeEvent.DATA_TYPE:
-        field.data_types = [value]
-    elif change == ChangeEvent.SOURCE:
-        node_type = "field" if "/" in value else "distribution"
-        source = mlc.Source(uid=value, node_type=node_type)
-        field.source = source
-    elif change == ChangeEvent.SOURCE_EXTRACT:
-        source = field.source
-        source = _get_source(source, value)
-        field.source = source
-    elif change == ChangeEvent.SOURCE_EXTRACT_COLUMN:
-        if not field.source:
-            field.source = mlc.Source(extract=mlc.Extract())
-        field.source.extract = mlc.Extract(column=value)
-    elif change == ChangeEvent.SOURCE_EXTRACT_JSON_PATH:
-        if not field.source:
-            field.source = mlc.Source(extract=mlc.Extract())
-        field.source.extract = mlc.Extract(json_path=value)
-    elif change == ChangeEvent.TRANSFORM:
-        number = kwargs.get("number")
-        if number is not None and number < len(field.source.transforms):
-            field.source.transforms[number] = mlc.Transform()
-    elif change == TransformType.FORMAT:
-        number = kwargs.get("number")
-        if number is not None and number < len(field.source.transforms):
-            field.source.transforms[number] = mlc.Transform(format=value)
-    elif change == TransformType.JSON_PATH:
-        number = kwargs.get("number")
-        if number is not None and number < len(field.source.transforms):
-            field.source.transforms[number] = mlc.Transform(json_path=value)
-    elif change == TransformType.REGEX:
-        number = kwargs.get("number")
-        if number is not None and number < len(field.source.transforms):
-            field.source.transforms[number] = mlc.Transform(regex=value)
-    elif change == TransformType.REPLACE:
-        number = kwargs.get("number")
-        if number is not None and number < len(field.source.transforms):
-            field.source.transforms[number] = mlc.Transform(replace=value)
-    elif change == TransformType.SEPARATOR:
-        number = kwargs.get("number")
-        if number is not None and number < len(field.source.transforms):
-            field.source.transforms[number] = mlc.Transform(separator=value)
-    elif change == ChangeEvent.REFERENCE:
-        node_type = "field" if "/" in value else "distribution"
-        source = mlc.Source(uid=value, node_type=node_type)
-        field.references = source
-    elif change == ChangeEvent.REFERENCE_EXTRACT:
-        source = field.references
-        source = _get_source(source, value)
-        field.references = source
-    elif change == ChangeEvent.REFERENCE_EXTRACT_COLUMN:
-        if not field.references:
-            field.references = mlc.Source(extract=mlc.Extract())
-        field.references.extract = mlc.Extract(column=value)
-    elif change == ChangeEvent.REFERENCE_EXTRACT_JSON_PATH:
-        if not field.references:
-            field.references = mlc.Source(extract=mlc.Extract())
-        field.references.extract = mlc.Extract(json_path=value)
-
-
 def render_source(
     record_set_key: int,
     record_set: RecordSet,
@@ -258,7 +126,7 @@ def render_source(
         options=[s for s in possible_sources if not s.startswith(record_set.name)],
         key=key,
         on_change=handle_field_change,
-        args=(ChangeEvent.SOURCE, field, key),
+        args=(FieldEvent.SOURCE, field, key),
     )
     if source.node_type == "distribution":
         extract = col2.selectbox(
@@ -267,7 +135,7 @@ def render_source(
             key=f"{prefix}-extract",
             options=EXTRACT_TYPES,
             on_change=handle_field_change,
-            args=(ChangeEvent.SOURCE_EXTRACT, field, key),
+            args=(FieldEvent.SOURCE_EXTRACT, field, key),
         )
         if extract == ExtractType.COLUMN:
             key = f"{prefix}-columnname"
@@ -276,7 +144,7 @@ def render_source(
                 value=source.extract.column,
                 key=key,
                 on_change=handle_field_change,
-                args=(ChangeEvent.SOURCE_EXTRACT_COLUMN, field, key),
+                args=(FieldEvent.SOURCE_EXTRACT_COLUMN, field, key),
             )
         if extract == ExtractType.JSON_PATH:
             key = f"{prefix}-jsonpath"
@@ -285,7 +153,7 @@ def render_source(
                 value=source.extract.json_path,
                 key=key,
                 on_change=handle_field_change,
-                args=(ChangeEvent.SOURCE_EXTRACT_JSON_PATH, field, key),
+                args=(FieldEvent.SOURCE_EXTRACT_JSON_PATH, field, key),
             )
 
     # Transforms
@@ -300,7 +168,7 @@ def render_source(
                 key=key,
                 options=TRANSFORM_TYPES,
                 on_change=handle_field_change,
-                args=(ChangeEvent.TRANSFORM, field, key),
+                args=(FieldEvent.TRANSFORM, field, key),
                 kwargs={"number": number},
             )
             if selected == TransformType.FORMAT:
@@ -404,7 +272,7 @@ def render_references(
             options=[s for s in possible_sources if not s.startswith(record_set.name)],
             key=key,
             on_change=handle_field_change,
-            args=(ChangeEvent.REFERENCE, field, key),
+            args=(FieldEvent.REFERENCE, field, key),
         )
         if references.node_type == "distribution":
             key = f"{key}-extract-references"
@@ -414,7 +282,7 @@ def render_references(
                 key=key,
                 options=EXTRACT_TYPES,
                 on_change=handle_field_change,
-                args=(ChangeEvent.REFERENCE_EXTRACT, field, key),
+                args=(FieldEvent.REFERENCE_EXTRACT, field, key),
             )
             if extract == ExtractType.COLUMN:
                 key = f"{key}-columnname"
@@ -423,7 +291,7 @@ def render_references(
                     value=references.extract.column,
                     key=key,
                     on_change=handle_field_change,
-                    args=(ChangeEvent.REFERENCE_EXTRACT_COLUMN, field, key),
+                    args=(FieldEvent.REFERENCE_EXTRACT_COLUMN, field, key),
                 )
             if extract == ExtractType.JSON_PATH:
                 key = f"{key}-jsonpath"
@@ -432,7 +300,7 @@ def render_references(
                     value=references.extract.json_path,
                     key=key,
                     on_change=handle_field_change,
-                    args=(ChangeEvent.REFERENCE_EXTRACT_JSON_PATH, field, key),
+                    args=(FieldEvent.REFERENCE_EXTRACT_JSON_PATH, field, key),
                 )
         col4.button(
             "✖️",
